requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
Deprecated==1.2.15
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
limits==4.0.1
packaging==24.2
pluggy==1.5.0
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.2
python-dotenv==1.0.1
redis==5.2.1
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
